ref: move tests for raw dataet to `raw` folder with respective files
